Split wrapper files by class
