+++ mobtimer-api/src/iWebSocketWrapper.ts
@@ -1,0 +1,9 @@
+
+export interface IWebSocketWrapper {
+  readyState: number;
+  send: (message: string) => void;
+  close: () => void;
+  OPEN: number;
+  CLOSED: number;
+  onmessage: (message: any) => void;
+}
